requirements.txt
streamlit
pandas
kagglehub
google-generativeai
matplotlib
seaborn
